# openapi/spec.yaml
openapi: 3.1.0
info:
  title: Pennant API
  version: 0.1.0
  summary: Feature flag / remote config API with real-time SDK push.
  description: |
    Pennant is a feature flag and remote configuration service. Define flags
    in a workspace, configure them per environment with targeting rules, and
    let SDKs evaluate them — either by hitting `/v1/evaluate` directly or by
    bootstrapping from `/v1/snapshot` and subscribing to `/v1/stream` for
    real-time updates.

    This spec is the source of truth. Laravel controllers conform to it; SDKs
    are generated from it. Run `npm run spectral` to validate locally.
  contact:
    name: Philip Rehberger
    url: https://pennant.philiprehberger.com
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.pennant.philiprehberger.com
    description: Production
  - url: http://localhost:8000
    description: Local development
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/flags/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      summary: Retrieve a flag
      operationId: getFlag
      tags: [Flags]
      responses:
        "200":
          description: The flag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flag"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update a flag
      operationId: updateFlag
      tags: [Flags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagInput"
      responses:
        "200":
          description: The updated flag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flag"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Archive a flag
      description: Soft-delete; flag stops being returned by SDK snapshots.
      operationId: archiveFlag
      tags: [Flags]
      responses:
        "204":
          description: Archived.
...
  /v1/evaluate:
    post:
      summary: Evaluate one or more flags against a context
      description: |
        Server-side evaluation. Pass a `context` and a list of `flags`; get
        back resolved values + reasons + the rule that matched.

        Use the snapshot endpoint instead if you can — local SDK evaluation
        is faster and cheaper, and behaves identically thanks to the
        cross-implementation corpus.
      operationId: evaluate
      tags: [Evaluation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateInput"
      responses:
        "200":
          description: Resolved flag values keyed by flag key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
      description: ULID.
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        `pn_srv_live_…` / `pn_srv_test_…` for server keys.
        `pn_clt_live_…` / `pn_clt_test_…` for client keys.
  responses:
...
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
      required: [title, status]
...
    Flag:
      type: object
      properties:
        id: { type: string }
        key: { type: string }
        type: { type: string, enum: [bool, string, number, json] }
        description: { type: string }
        default_value: {}
        archived_at: { type: string, format: date-time, nullable: true }
        configurations:
          type: array
          items: { $ref: "#/components/schemas/FlagConfiguration" }
        created_at: { type: string, format: date-time }
      required: [id, key, type, default_value]
    FlagInput:
      type: object
      properties:
        key:
          type: string
          minLength: 1
          maxLength: 80
          pattern: "^[a-z0-9][a-z0-9_-]*$"
        type: { type: string, enum: [bool, string, number, json] }
        description: { type: string, maxLength: 500 }
        default_value: {}
      required: [key, type, default_value]
...
    FlagConfiguration:
      type: object
      properties:
        id: { type: string }
        flag_id: { type: string }
        environment_key: { type: string }
        state: { type: string, enum: [on, off] }
        variation: {}
        bucketing_attribute: { type: string }
        bucketing_seed: { type: string }
        rules:
          type: array
          items: { $ref: "#/components/schemas/TargetingRule" }
        updated_at: { type: string, format: date-time }
      required: [id, flag_id, environment_key, state, variation, rules]
...
    TargetingRule:
      type: object
      properties:
        id: { type: string }
        priority: { type: integer }
        description: { type: string }
        condition: { $ref: "#/components/schemas/RuleExpression" }
        variation: {}
      required: [priority, condition, variation]
...
    RuleExpression:
      description: |
        A JSON expression tree. Supported forms:

        - `{ attribute, op, value }` — terminal comparison.
        - `{ segment }` — references a segment by key.
        - `{ all: [expr, …] }` — AND.
        - `{ any: [expr, …] }` — OR.
        - `{ none: [expr, …] }` — NOT (matches if no child matches).

        Operators (terminal `op`):
        equals, not_equals, in, not_in, contains, not_contains,
        starts_with, ends_with, regex, gt, gte, lt, lte, before, after,
        percentage.

        The `percentage` operator uses the flag configuration's
        `bucketing_attribute` + `bucketing_seed` to deterministically place
        the context in `[0, 1)`; rule matches when bucket < value / 100.
      oneOf:
        - type: object
          properties:
            attribute: { type: string }
            op:
              type: string
              enum:
                [
                  equals,
                  not_equals,
                  in,
                  not_in,
                  contains,
                  not_contains,
                  starts_with,
                  ends_with,
                  regex,
                  gt,
                  gte,
                  lt,
                  lte,
                  before,
                  after,
                  percentage,
                ]
            value: {}
          required: [attribute, op, value]
        - type: object
          properties:
            segment: { type: string }
          required: [segment]
        - type: object
          properties:
            all:
              type: array
              items: { $ref: "#/components/schemas/RuleExpression" }
          required: [all]
        - type: object
          properties:
            any:
              type: array
              items: { $ref: "#/components/schemas/RuleExpression" }
          required: [any]
        - type: object
          properties:
            none:
              type: array
              items: { $ref: "#/components/schemas/RuleExpression" }
          required: [none]
...
    EvaluateInput:
      type: object
      properties:
        environment: { type: string }
        context:
          type: object
          additionalProperties: true
          description: |
            User attributes used for rule evaluation. By convention
            includes `userId`; other keys are free-form (`plan`, `country`,
            `email`, …).
        flags:
          type: array
          items: { type: string }
          description: Flag keys to evaluate. Omit to evaluate all.
      required: [environment, context]
    EvaluateResult:
      type: object
      additionalProperties:
        type: object
        properties:
          value: {}
          reason:
            type: string
            enum: [default, off, rule_match, percentage_rollout, fallthrough]
          rule_id: { type: string, nullable: true }
        required: [value, reason]
...